Consolidate core services into shared module
