+++ client/app/src/app/app.component.ts
@@ -2,9 +2,9 @@ import { Component, OnInit, QueryList, ViewChild } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { MatSidenav } from '@angular/material';
 import { Router } from '@angular/router';
-import { AuthenticationService } from './core/auth/_services/authentication.service';
-import { UserService } from './core/auth/_services/user.service';
-import { WindowRef } from './core/window-ref.service';
+import { AuthenticationService } from './shared/_services/authentication.service';
+import { UserService } from './shared/_services/user.service';
+import { WindowRef } from './shared/_services/window-ref.service';
 import { updates } from './core/update/update/updates';
 import { TangyFormService } from './tangy-forms/tangy-form-service';
 import PouchDB from 'pouchdb';
