+++ client/app/src/app/case-management/case-details/case-details.component.spec.ts
@@ -2,13 +2,33 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CaseDetailsComponent } from './case-details.component';
 
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { CaseManagementModule } from '../case-management.module';
+import { AppRoutingModule } from 'src/app/app-routing.module';
+export function HttpClientLoaderFactory(httpClient: HttpClient) {
+  return new TranslateHttpLoader(httpClient, './assets/', '.json');
+}
+
 describe('CaseDetailsComponent', () => {
   let component: CaseDetailsComponent;
   let fixture: ComponentFixture<CaseDetailsComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CaseDetailsComponent ]
+      imports: [ 
+        CaseManagementModule,
+        AppRoutingModule,
+        TranslateModule.forRoot({
+          loader: {
+            provide: TranslateLoader,
+            useFactory: HttpClientLoaderFactory,
+            deps: [HttpClient]
+          }
+        })
+      ]
+
     })
     .compileComponents();
   }));
