+++ client/app/src/app/case/components/event-form/event-form.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild, ElementRef, AfterContentInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { WindowRef } from '../../../core/window-ref.service';
-import { UserService } from '../../../core/auth/_services/user.service';
+import { WindowRef } from '../../../shared/_services/window-ref.service';
+import { UserService } from '../../../shared/_services/user.service';
 import { CaseService } from '../../services/case.service'
 import { EventForm } from '../../classes/event-form.class';
 import { CaseEvent } from '../../classes/case-event.class';
