+++ client/app/src/app/class/_services/dashboard.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import PouchDB from 'pouchdb';
-import {AuthenticationService} from "../../core/auth/_services/authentication.service";
+import {AuthenticationService} from "../../shared/_services/authentication.service";
 import {FormMetadata} from "../form-metadata";
 import {StudentResult} from "../reports/student-grouping-report/student-result";
 import {Feedback} from "../feedback";
