+++ client/app/src/app/class/dashboard/dashboard.component.ts
@@ -2,7 +2,7 @@ import {Component, ElementRef, OnInit, ViewChild} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {DashboardService} from "../_services/dashboard.service";
 import {MatTableDataSource, PageEvent} from "@angular/material";
-import {AuthenticationService} from "../../core/auth/_services/authentication.service";
+import {AuthenticationService} from "../../shared/_services/authentication.service";
 import {ClassFormService} from "../_services/class-form.service";
 import {Router} from "@angular/router";
 import {_TRANSLATE} from "../../shared/translation-marker";
