+++ client/app/src/app/class/reports/student-subtest-report/student-subtest-report.component.ts
@@ -2,7 +2,7 @@ import {Component, ElementRef, OnInit, ViewChild} from '@angular/core';
 import {ActivatedRoute} from "@angular/router";
 import {ClassFormService} from "../../_services/class-form.service";
 import {DashboardService} from "../../_services/dashboard.service";
-import {AuthenticationService} from "../../../core/auth/_services/authentication.service";
+import {AuthenticationService} from "../../../shared/_services/authentication.service";
 import {ClassUtils} from "../../class-utils";
 import {AppConfigService} from "../../../shared/_services/app-config.service";
 import {StudentResult} from "../student-grouping-report/student-result";
