+++ client/app/src/app/core/auth/_guards/login-guard.service.spec.ts
@@ -1,8 +1,8 @@
 import { TestBed, inject, async } from '@angular/core/testing';
 import { Router } from '@angular/router';
 import { LoginGuard } from './login-guard.service';
-import { AuthenticationService } from '../_services/authentication.service';
-import { UserService } from '../_services/user.service';
+import { AuthenticationService } from '../../../shared/_services/authentication.service';
+import { UserService } from '../../../shared/_services/user.service';
 
 describe('LoginGuard', () => {
   const router = { navigate: jasmine.createSpy('navigate') };
