+++ client/app/src/app/core/auth/auth.module.ts
@@ -5,8 +5,6 @@ import { MatButtonModule, MatInputModule, MatSelectModule } from '@angular/mater
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { LoginGuard } from './_guards/login-guard.service';
-import { AuthenticationService } from './_services/authentication.service';
-import { UserService } from './_services/user.service';
 import { AuthRoutingModule } from './auth-routing.module';
 import { LoginComponent } from './login/login.component';
 import { RegistrationComponent } from './registration/registration.component';
@@ -25,7 +23,7 @@ import { SharedModule } from '../../shared/shared.module';
     AuthRoutingModule,
     SharedModule
   ],
-  providers: [LoginGuard, AuthenticationService, UserService],
+  providers: [LoginGuard],
   declarations: [LoginComponent, RegistrationComponent]
 })
 export class AuthModule { }
