+++ client/app/src/app/core/auth/registration/registration.component.ts
@@ -6,9 +6,9 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AppConfigService } from '../../../shared/_services/app-config.service';
 
-import { AuthenticationService } from '../_services/authentication.service';
-import { UserService } from '../_services/user.service';
-import { User } from './../_services/user.model.interface';
+import { AuthenticationService } from '../../../shared/_services/authentication.service';
+import { UserService } from '../../../shared/_services/user.service';
+import { User } from '../../../shared/_services/user.model.interface';
 import { _TRANSLATE } from '../../../shared/translation-marker';
