+++ client/app/src/app/shared/_services/authentication.service.spec.ts
