+++ client/app/src/app/shared/_services/user.model.interface.ts
