+++ client/app/src/app/shared/_services/window-ref.service.ts
