+++ client/app/src/app/user-profile/user-profile.component.ts
@@ -2,7 +2,7 @@ import { AfterContentInit, ElementRef, Component, ViewChild } from '@angular/cor
 import { ActivatedRoute, Router } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 
-import { UserService } from '../core/auth/_services/user.service';
+import { UserService } from '../shared/_services/user.service';
 import { TangyFormService } from '../tangy-forms/tangy-form-service';
 import {AppConfigService} from "../shared/_services/app-config.service";
